Zero copy tests passing for JMAP

+++ crates/smtp/src/outbound/delivery.rs
@@ -4,22 +4,22 @@
  * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-SEL
  */
 
-use crate::outbound::client::{from_error_status, from_mail_send_error, SmtpClient};
+use crate::outbound::client::{SmtpClient, from_error_status, from_mail_send_error};
 use crate::outbound::dane::dnssec::TlsaLookup;
 use crate::outbound::lookup::DnsLookup;
 use crate::outbound::mta_sts::lookup::MtaStsLookup;
 use crate::outbound::mta_sts::verify::VerifyPolicy;
 use crate::outbound::{client::StartTlsResult, dane::verify::TlsaVerify};
 use crate::queue::dsn::SendDsn;
-use crate::queue::spool::{SmtpSpool, LOCK_EXPIRY};
+use crate::queue::spool::{LOCK_EXPIRY, SmtpSpool};
 use crate::queue::throttle::IsAllowed;
 use crate::reporting::SmtpReporting;
+use common::Server;
 use common::config::{
     server::ServerProtocol,
     smtp::{queue::RequireOptional, report::AggregateFrequency},
 };
 use common::ipc::{PolicyType, QueueEvent, QueueEventStatus, TlsEvent};
-use common::Server;
 use mail_auth::{
     mta_sts::TlsRpt,
     report::tlsrpt::{FailureDetails, ResultType},
@@ -30,7 +30,7 @@ use std::{
     net::{IpAddr, Ipv4Addr, SocketAddr},
     time::{Duration, Instant},
 };
-use store::write::{now, BatchBuilder, QueueClass, ValueClass};
+use store::write::{BatchBuilder, QueueClass, ValueClass, now};
 use trc::{DaneEvent, DeliveryEvent, MtaStsEvent, ServerEvent, TlsRptEvent};
 
 use crate::{
@@ -38,7 +38,7 @@ use crate::{
     reporting::tls::TlsRptOptions,
 };
 
-use super::{lookup::ToNextHop, mta_sts, session::SessionParams, NextHop, TlsStrategy};
+use super::{NextHop, TlsStrategy, lookup::ToNextHop, mta_sts, session::SessionParams};
 use crate::queue::{Domain, Error, QueueEnvelope, QueuedMessage, Status};
 
 impl QueuedMessage {
@@ -104,9 +104,10 @@ impl QueuedMessage {
                     )));
 
                     if let Err(err) = server.store().write(batch.build()).await {
-                        trc::error!(err
-                            .details("Failed to delete queue event.")
-                            .caused_by(trc::location!()));
+                        trc::error!(
+                            err.details("Failed to delete queue event.")
+                                .caused_by(trc::location!())
+                        );
                     }
 
                     // Unlock event
@@ -243,7 +244,9 @@ impl QueuedMessage {
                     // Deliver message locally
                     let delivery_result = message
                         .deliver_local(
-                            recipients.iter_mut().filter(|r| r.domain_idx == domain_idx),
+                            recipients
+                                .iter_mut()
+                                .filter(|r| r.domain_idx == domain_idx as u32),
                             &server,
                         )
                         .await;
@@ -1096,7 +1099,7 @@ impl QueuedMessage {
                                             smtp_client,
                                             recipients
                                                 .iter_mut()
-                                                .filter(|r| r.domain_idx == domain_idx),
+                                                .filter(|r| r.domain_idx == domain_idx as u32),
                                             params,
                                         )
                                         .await
@@ -1154,7 +1157,7 @@ impl QueuedMessage {
                                                 smtp_client,
                                                 recipients
                                                     .iter_mut()
-                                                    .filter(|r| r.domain_idx == domain_idx),
+                                                    .filter(|r| r.domain_idx == domain_idx as u32),
                                                 params,
                                             )
                                             .await
@@ -1211,7 +1214,9 @@ impl QueuedMessage {
                             message
                                 .deliver(
                                     smtp_client,
-                                    recipients.iter_mut().filter(|r| r.domain_idx == domain_idx),
+                                    recipients
+                                        .iter_mut()
+                                        .filter(|r| r.domain_idx == domain_idx as u32),
                                     params,
                                 )
                                 .await
@@ -1261,7 +1266,9 @@ impl QueuedMessage {
                         message
                             .deliver(
                                 smtp_client,
-                                recipients.iter_mut().filter(|r| r.domain_idx == domain_idx),
+                                recipients
+                                    .iter_mut()
+                                    .filter(|r| r.domain_idx == domain_idx as u32),
                                 params,
                             )
                             .await
@@ -1341,7 +1348,7 @@ impl Message {
                     );
 
                     for rcpt in &mut self.recipients {
-                        if rcpt.domain_idx == idx {
+                        if rcpt.domain_idx == idx as u32 {
                             rcpt.status = std::mem::replace(&mut rcpt.status, Status::Scheduled)
                                 .into_permanent();
                         }
@@ -1359,7 +1366,7 @@ impl Message {
                     );
 
                     for rcpt in &mut self.recipients {
-                        if rcpt.domain_idx == idx {
+                        if rcpt.domain_idx == idx as u32 {
                             rcpt.status = std::mem::replace(&mut rcpt.status, Status::Scheduled)
                                 .into_permanent();
                         }
